refactor(filesystem): extract tests into tests/ sub-module (#40)
Move the #[cfg(test)] mod tests { ... } block out of commands.rs into
a new filesystem/tests/ sub-module with one file per command:
  - tests/mod.rs     — shared home_test_dir helper + re-exports
  - tests/list_tests.rs  — 3 list_dir tests
  - tests/read_tests.rs  — 4 read_file tests
  - tests/write_tests.rs — 8 write_file tests

Production code in commands.rs is unchanged. This drops commands.rs
from 864 → ~489 lines, getting the file back under the 800-line
CLAUDE.md guideline.

Test count: 15 before, 15 after. All regression tests from PR #38
pass under the new module paths.
