+++ src-tauri/src/filesystem/commands.rs
@@ -486,379 +486,3 @@ pub fn write_file(request: WriteFileRequest) -> Result<(), String> {
 
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-
-    /// Create a temp dir under $HOME so it passes the home-directory scope check
-    fn home_test_dir(name: &str) -> std::path::PathBuf {
-        dirs::home_dir()
-            .expect("HOME must be set for tests")
-            .join(format!(".vimeflow_test_{}", name))
-    }
-
-    #[test]
-    fn list_dir_returns_sorted_entries() {
-        let dir = home_test_dir("list_dir");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-        fs::create_dir(dir.join("beta")).unwrap();
-        fs::create_dir(dir.join("alpha")).unwrap();
-        fs::write(dir.join("zebra.txt"), "").unwrap();
-        fs::write(dir.join("apple.txt"), "").unwrap();
-
-        let result = list_dir(ListDirRequest {
-            path: dir.to_string_lossy().to_string(),
-        });
-
-        assert!(result.is_ok());
-        let entries = result.unwrap();
-        assert_eq!(entries.len(), 4);
-        // Folders first, sorted
-        assert_eq!(entries[0].name, "alpha");
-        assert_eq!(entries[0].entry_type, EntryType::Folder);
-        assert_eq!(entries[1].name, "beta");
-        assert_eq!(entries[1].entry_type, EntryType::Folder);
-        // Files second, sorted
-        assert_eq!(entries[2].name, "apple.txt");
-        assert_eq!(entries[2].entry_type, EntryType::File);
-        assert_eq!(entries[3].name, "zebra.txt");
-        assert_eq!(entries[3].entry_type, EntryType::File);
-
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn list_dir_skips_hidden_files() {
-        let dir = home_test_dir("hidden");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-        fs::write(dir.join(".hidden"), "").unwrap();
-        fs::write(dir.join("visible.txt"), "").unwrap();
-
-        let result = list_dir(ListDirRequest {
-            path: dir.to_string_lossy().to_string(),
-        });
-
-        assert!(result.is_ok());
-        let entries = result.unwrap();
-        assert_eq!(entries.len(), 1);
-        assert_eq!(entries[0].name, "visible.txt");
-
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn list_dir_rejects_nonexistent_path() {
-        let result = list_dir(ListDirRequest {
-            path: "/nonexistent/path/abc123".to_string(),
-        });
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn read_file_returns_content() {
-        let dir = home_test_dir("read_file");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-        fs::write(dir.join("hello.txt"), "hello world").unwrap();
-
-        let result = read_file(ReadFileRequest {
-            path: dir.join("hello.txt").to_string_lossy().to_string(),
-        });
-
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), "hello world");
-
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn read_file_rejects_path_outside_home() {
-        let result = read_file(ReadFileRequest {
-            path: "/etc/passwd".to_string(),
-        });
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("access denied"));
-    }
-
-    #[test]
-    fn read_file_rejects_directory() {
-        let dir = home_test_dir("read_file_dir");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-
-        let result = read_file(ReadFileRequest {
-            path: dir.to_string_lossy().to_string(),
-        });
-
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("not a file"));
-
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn write_file_creates_file() {
-        let dir = home_test_dir("write_file");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-
-        let file_path = dir.join("new_file.txt");
-        let result = write_file(WriteFileRequest {
-            path: file_path.to_string_lossy().to_string(),
-            content: "test content".to_string(),
-        });
-
-        assert!(result.is_ok());
-        assert!(file_path.exists());
-        let content = fs::read_to_string(&file_path).unwrap();
-        assert_eq!(content, "test content");
-
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn write_file_creates_parent_dirs() {
-        let dir = home_test_dir("write_file_nested");
-        let _ = fs::remove_dir_all(&dir);
-
-        let file_path = dir.join("subdir").join("nested").join("file.txt");
-        let result = write_file(WriteFileRequest {
-            path: file_path.to_string_lossy().to_string(),
-            content: "nested content".to_string(),
-        });
-
-        assert!(result.is_ok());
-        assert!(file_path.exists());
-        let content = fs::read_to_string(&file_path).unwrap();
-        assert_eq!(content, "nested content");
-
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn write_file_rejects_path_outside_home() {
-        let result = write_file(WriteFileRequest {
-            path: "/etc/test_forbidden.txt".to_string(),
-            content: "forbidden".to_string(),
-        });
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("access denied"));
-    }
-
-    #[test]
-    fn write_file_rejects_traversal_into_sibling_of_home() {
-        // A path like `~/../etc/evil.txt` must be rejected WITHOUT creating
-        // any directories on disk (the P1 fix) — the previous implementation
-        // ran `create_dir_all` before canonicalizing, which could mutate the
-        // filesystem outside the home scope.
-        let home = dirs::home_dir().expect("HOME must be set for tests");
-        let home_parent = home
-            .parent()
-            .expect("home dir should have a parent in tests");
-
-        // Forge a path that escapes home via `..`
-        let evil = home.join("..").join(".vimeflow_traversal_test").join("evil.txt");
-        let marker = home_parent.join(".vimeflow_traversal_test");
-        let _ = fs::remove_dir_all(&marker);
-
-        let result = write_file(WriteFileRequest {
-            path: evil.to_string_lossy().to_string(),
-            content: "should be rejected".to_string(),
-        });
-
-        assert!(result.is_err(), "traversal path must be rejected");
-        assert!(result.unwrap_err().contains("access denied"));
-
-        // Crucial: verify that NO directory was created outside home.
-        assert!(
-            !marker.exists(),
-            "traversal fix must not create directories outside home: {}",
-            marker.display()
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn read_file_refuses_to_follow_symlink_escaping_home() {
-        // Mirror of the write_file symlink regression test. If a symlink
-        // inside home points outside home and `read_file` follows it,
-        // the webview would receive sandbox-escaped contents.
-        use std::os::unix::fs::symlink;
-
-        let dir = home_test_dir("read_file_symlink_escape");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-
-        // Point the symlink at a real file outside home so canonicalize() resolves
-        // and the scope check would otherwise accept it if not for O_NOFOLLOW.
-        let outside_target = std::env::temp_dir().join(
-            ".vimeflow_read_file_symlink_secret.txt",
-        );
-        fs::write(&outside_target, "SECRET").unwrap();
-
-        let link = dir.join("innocent_name.txt");
-        symlink(&outside_target, &link).unwrap();
-
-        // The expand_home+canonicalize path will resolve the symlink and
-        // realize it's outside home, so the scope check rejects it here.
-        // (The O_NOFOLLOW guard is defense-in-depth for the TOCTOU race
-        // where the symlink is introduced between canonicalize and open.)
-        let result = read_file(ReadFileRequest {
-            path: link.to_string_lossy().to_string(),
-        });
-
-        assert!(result.is_err(), "symlink-escape read must be rejected");
-        let err = result.unwrap_err();
-        assert!(
-            err.contains("access denied") || err.contains("invalid path"),
-            "expected scope rejection, got: {}",
-            err
-        );
-
-        let _ = fs::remove_file(&outside_target);
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn write_file_refuses_to_follow_symlink_escaping_home() {
-        // A symlink inside home pointing outside home must not let `fs::write`
-        // escape the sandbox. The previous implementation only canonicalized
-        // the parent directory, so `fs::write` would follow the target
-        // symlink and mutate files anywhere on disk the process could reach.
-        use std::os::unix::fs::symlink;
-
-        let dir = home_test_dir("write_file_symlink_escape");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-
-        // The symlink target is a path outside home. We never want to touch it.
-        let outside_target = std::env::temp_dir().join(
-            ".vimeflow_write_file_symlink_escape_target.txt",
-        );
-        let _ = fs::remove_file(&outside_target);
-
-        let link = dir.join("evil_link");
-        // evil_link -> /tmp/.vimeflow_write_file_symlink_escape_target.txt
-        symlink(&outside_target, &link).unwrap();
-
-        let result = write_file(WriteFileRequest {
-            path: link.to_string_lossy().to_string(),
-            content: "should not escape".to_string(),
-        });
-
-        assert!(result.is_err(), "symlink write must be rejected");
-        assert!(result.unwrap_err().contains("access denied"));
-
-        // Crucial: verify nothing was written outside home.
-        assert!(
-            !outside_target.exists(),
-            "symlink guard must not write outside home: {}",
-            outside_target.display()
-        );
-
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn write_file_refuses_intermediate_symlink_escape() {
-        // Cover the TOCTOU class where an INTERMEDIATE path component is
-        // swapped for a symlink pointing outside home. O_NOFOLLOW on the
-        // final open() only guards the last segment, so this must be
-        // caught by re-canonicalizing the parent after create_dir_all.
-        use std::os::unix::fs::symlink;
-
-        let dir = home_test_dir("write_file_intermediate_symlink");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-
-        // Pre-plant a symlink at an "intermediate" directory that points
-        // outside home. The target path uses this link as a parent.
-        let outside = std::env::temp_dir().join(".vimeflow_intermediate_escape");
-        let _ = fs::remove_dir_all(&outside);
-        fs::create_dir_all(&outside).unwrap();
-
-        let link = dir.join("escape_link");
-        symlink(&outside, &link).unwrap();
-
-        // raw path: dir/escape_link/file.txt
-        //   - walk-up check: dir/escape_link exists → canonicalize resolves
-        //     it to the outside path → starts_with(home) fails → rejected.
-        let evil = link.join("file.txt");
-
-        let result = write_file(WriteFileRequest {
-            path: evil.to_string_lossy().to_string(),
-            content: "should not escape".to_string(),
-        });
-
-        assert!(result.is_err(), "intermediate symlink write must be rejected");
-        assert!(result.unwrap_err().contains("access denied"));
-
-        // Confirm the attacker target is untouched.
-        assert!(
-            !outside.join("file.txt").exists(),
-            "intermediate symlink guard must not write outside home"
-        );
-
-        let _ = fs::remove_dir_all(&outside);
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn write_file_refuses_symlink_even_to_in_home_target() {
-        // Stricter: reject *any* symlink at the target position, even ones
-        // pointing inside home. This closes the TOCTOU window where a symlink
-        // could be swapped between our check and the write.
-        use std::os::unix::fs::symlink;
-
-        let dir = home_test_dir("write_file_symlink_inner");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-
-        let real_target = dir.join("real.txt");
-        fs::write(&real_target, "original").unwrap();
-
-        let link = dir.join("link.txt");
-        symlink(&real_target, &link).unwrap();
-
-        let result = write_file(WriteFileRequest {
-            path: link.to_string_lossy().to_string(),
-            content: "nope".to_string(),
-        });
-
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("symlink"));
-
-        // Real target should be untouched.
-        assert_eq!(fs::read_to_string(&real_target).unwrap(), "original");
-
-        let _ = fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn write_file_overwrites_existing() {
-        let dir = home_test_dir("write_file_overwrite");
-        let _ = fs::remove_dir_all(&dir);
-        fs::create_dir_all(&dir).unwrap();
-
-        let file_path = dir.join("overwrite.txt");
-        fs::write(&file_path, "original").unwrap();
-
-        let result = write_file(WriteFileRequest {
-            path: file_path.to_string_lossy().to_string(),
-            content: "updated".to_string(),
-        });
-
-        assert!(result.is_ok());
-        let content = fs::read_to_string(&file_path).unwrap();
-        assert_eq!(content, "updated");
-
-        let _ = fs::remove_dir_all(&dir);
-    }
-}
