+++ src-tauri/src/filesystem/mod.rs
@@ -2,3 +2,6 @@ mod commands;
 mod types;
 
 pub use commands::{list_dir, read_file, write_file};
+
+#[cfg(test)]
+mod tests;
